@@ -9,16 +9,15 @@ classifiers =
 	Topic :: System :: Archiving :: Packaging
 	License :: OSI Approved :: MIT License
 	Programming Language :: Python
-	Programming Language :: Python :: 2
-	Programming Language :: Python :: 2.7
-	Programming Language :: Python :: 3
-	Programming Language :: Python :: 3.5
-	Programming Language :: Python :: 3.6
+	Programming Language :: Python :: 3 :: Only
 	Programming Language :: Python :: 3.7
 	Programming Language :: Python :: 3.8
+	Programming Language :: Python :: 3.9
+	Programming Language :: Python :: 3.10
+	Programming Language :: Python :: 3.11
 author = Daniel Holth
 author_email = dholth@fastmail.fm
-maintainer = Alex Gronholm
+maintainer = Alex Grönholm
 maintainer_email = alex.gronholm@nextday.fi
 url = https://github.com/pypa/wheel
 project_urls = 
@@ -32,8 +31,7 @@ license = MIT
 package_dir = 
 	= src
 packages = find:
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*
-setup_requires = setuptools >= 40.9.0
+python_requires = >=3.7
 zip_safe = False
 
 [options.packages.find]
@@ -42,7 +40,6 @@ where = src
 [options.extras_require]
 test = 
 	pytest >= 3.0.0
-	pytest-cov
 
 [options.entry_points]
 console_scripts = 
@@ -50,21 +47,23 @@ console_scripts =
 distutils.commands = 
 	bdist_wheel = wheel.bdist_wheel:bdist_wheel
 
+[tool:isort]
+src_paths = src
+profile = black
+skip_gitignore = true
+
+[flake8]
+max-line-length = 88
+
 [tool:pytest]
-addopts = --cov=wheel
 testpaths = tests
 
 [coverage:run]
 source = wheel
+omit = */vendored/*
 
 [coverage:report]
 show_missing = true
-
-[flake8]
-max-line-length = 99
-
-[bdist_wheel]
-universal = 1
 
 [egg_info]
 tag_build = 
