@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import os.path
 import re
 
@@ -7,7 +9,7 @@ from wheel.wheelfile import WHEEL_INFO_RE
 
 def test_egg_re():
     """Make sure egg_info_re matches."""
-    egg_names_path = os.path.join(os.path.dirname(__file__), 'eggnames.txt')
+    egg_names_path = os.path.join(os.path.dirname(__file__), "eggnames.txt")
     with open(egg_names_path) as egg_names:
         for line in egg_names:
             line = line.strip()
@@ -20,5 +22,7 @@ def test_convert_egg(egg_paths, tmpdir):
     wheel_names = [path.basename for path in tmpdir.listdir()]
     assert len(wheel_names) == len(egg_paths)
     assert all(WHEEL_INFO_RE.match(filename) for filename in wheel_names)
-    assert all(re.match(r'^[\w\d.]+-\d\.\d-\w+\d+-[\w\d]+-[\w\d]+\.whl$', fname)
-               for fname in wheel_names)
+    assert all(
+        re.match(r"^[\w\d.]+-\d\.\d-\w+\d+-[\w\d]+-[\w\d]+\.whl$", fname)
+        for fname in wheel_names
+    )
